* Added charge meter to the UI, can be disabled via GUI * Trample and Lucifer Cannon charging reworked and moved to PMove * Can no longer "cancel" a charging Lucifer Cannon * STAT_MISC2 is now unused, booster and charge information can be inferred elsewhere (STAT_UNUSED) * Trying to fire an empty weapon makes a clicking noise (also a bugfix, apparently the server would spam EV_NOAMMO which did nothing but is now used for the clicking noise -- audible to all players) * Created an alternate, muffled Lucifer Cannon warning noise for other people's Lucifer Cannons * Fixed bug which prevented players from switching to a different player while spectating someone in the spawn queue * Spectators are now properly moved to the lock view position when spectating a player in the spawn queue

+++ src/cgame/cg_view.c
@@ -471,7 +471,7 @@ static void CG_OffsetFirstPersonView( void )
     if( ps->stats[ STAT_MISC ] > 0 )
     {
       float fraction = (float)ps->stats[ STAT_MISC ] /
-        (float)LEVEL4_TRAMPLE_CHARGE_MAX;
+                       LEVEL4_TRAMPLE_CHARGE_MAX;
 
       if( fraction > 1.0f )
         fraction = 1.0f;
